ruff fix

+++ bot/helper/mirror_leech_utils/gdrive_utils/upload.py
@@ -47,6 +47,9 @@ class GoogleDriveUpload(GoogleDriveHelper):
         self.service = self.authorize()
         LOGGER.info(f"Uploading: {self._path}")
         self._updater = SetInterval(self.update_interval, self.progress)
+        link = None
+        dir_id = None
+        mime_type = None
         try:
             if ospath.isfile(self._path):
                 if self._path.lower().endswith(
@@ -96,24 +99,25 @@ class GoogleDriveUpload(GoogleDriveHelper):
             self._is_errored = True
         finally:
             self._updater.cancel()
-            if self.listener.is_cancelled and not self._is_errored:
-                if mime_type == "Folder":
-                    LOGGER.info("Deleting uploaded data from Drive...")
-                    self.service.files().delete(
-                        fileId=dir_id,
-                        supportsAllDrives=True,
-                    ).execute()
-                return
-            if self._is_errored:
-                return
-            async_to_sync(
-                self.listener.on_upload_complete,
-                link,
-                self.total_files,
-                self.total_folders,
-                mime_type,
-                dir_id=self.get_id_from_url(link),
-            )
+    
+        if self.listener.is_cancelled and not self._is_errored:
+            if mime_type == "Folder" and dir_id:
+                LOGGER.info("Deleting uploaded data from Drive...")
+                self.service.files().delete(
+                    fileId=dir_id,
+                    supportsAllDrives=True,
+                ).execute()
+            return
+        if self._is_errored:
+            return
+        async_to_sync(
+            self.listener.on_upload_complete,
+            link,
+            self.total_files,
+            self.total_folders,
+            mime_type,
+            dir_id=self.get_id_from_url(link) if link else None,
+        )
 
     def _upload_dir(self, input_directory, dest_id, unwanted_files, ft_delete):
         list_dirs = listdir(input_directory)
